+++ packages/hono-takibi/src/core/helper/get-camel-case-schema-name-helper.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { getCamelCaseSchemaName } from './get-camel-case-schema-name'
+import { getCamelCaseSchemaName } from './get-camel-case-schema-name-helper'
 
 const getCamelCaseSchemaNameTestCases = [
   {
