+++ packages/hono-takibi/src/core/helper/get-camel-case-schema-name-helper.ts
@@ -1,4 +1,4 @@
-import { decapitalize } from '../../text/decapitalize'
+import { decapitalize } from '../text/decapitalize'
 
 /**
  * Generates a camelCase schema name from a given schema name.
