+++ packages/hono-takibi/src/core/helper/get-pascal-case-schema-name-helper.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { getPascalCaseSchemaName } from './get-pascal-case-schema-name'
+import { getPascalCaseSchemaName } from './get-pascal-case-schema-name-helper'
 
 const getPascalCaseSchemaNameTestCases = [
   {
