+++ packages/hono-takibi/src/core/helper/get-pascal-case-schema-name-helper.ts
@@ -1,4 +1,4 @@
-import { capitalize } from '../../text/capitalize'
+import { capitalize } from '../text/capitalize'
 
 /**
  * Generates a PascalCase schema name from a given schema name.
