+++ packages/hono-takibi/src/core/helper/get-variable-schema-name-helper.ts
@@ -1,6 +1,6 @@
 import type { Config } from '../../config'
-import { getCamelCaseSchemaName } from '../schema/references/get-camel-case-schema-name'
-import { getPascalCaseSchemaName } from '../schema/references/get-pascal-case-schema-name'
+import { getCamelCaseSchemaName } from './get-camel-case-schema-name-helper'
+import { getPascalCaseSchemaName } from './get-pascal-case-schema-name-helper'
 
 export const getVariableSchemaNameHelper = (name: string, config: Config) => {
   return config.schema.name === 'camelCase'
